pyproject.toml
[project]
name = "gitlab-release-notes"
version = "0.1.0"
description = "A LangGraph-powered application that automatically generates comprehensive release notes by analyzing GitLab activity"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "python-gitlab>=3.15.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "openai>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true